Refactor: 소켓 관련 코드 service 파일로 이동
